+++ Spidy/Assets/Scripts/LeftOrRight.cs
@@ -1,0 +1,25 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using System;
+
+public class LeftOrRight : MonoBehaviour
+{
+    void PutWebLeftOrRightRandomly (object sender, Spawn.OnRespawnArgs e) 
+    {
+        e.webTransform.position += new Vector3 (LeftOrRightFromCenter(),0,0);
+    }
+
+    int LeftOrRightFromCenter ()
+    {
+        if ( UnityEngine.Random.Range(0, 2) != 0) return 1 ;
+        return -1;
+    }
+    void Start()
+    {
+        Spawn webSpawner = GetComponent<Spawn>();
+        webSpawner.OnRespawn += PutWebLeftOrRightRandomly; 
+    }
+
+    
+}